fixed profile pic delete

+++ views/v_users_profile.php
@@ -45,7 +45,9 @@
 		<?php if($avatars[$i] != ""): ?>
 			<img class="img_big" src="<?=$avatars[$i]?>">
 			<br>
-			<a class="smalllink" href="/users/deletepic/<?=$i?>">Delete Photo</a>
+			<?php if($this_user['user_id'] == $user->user_id)?>
+				<a class="smalllink" href="/users/deletepic/<?=$i?>">Delete Photo</a>
+			<?php endif; ?>
 			<br><br>
 		<?php endif; ?>
 	<?php endfor; ?>
